Pin clojure build image to an older tag
With clojure base image tagged as "latest", it is not possible to build that
image. The following error is returned:

  Step 6/15 : RUN mv "$(lein uberjar | sed -n 's/^Created \(.*standalone\.jar\)/\1/p')" myapp-standalone.jar
   ---> Running in 288116953c6e
  Compiling stencil.service.core
  java.lang.ClassNotFoundException: javax.xml.bind.DatatypeConverter, compiling:(server.clj:1:1)
  Exception in thread "main" java.lang.ClassNotFoundException: javax.xml.bind.DatatypeConverter, compiling:(server.clj:1:1)
          at clojure.lang.Compiler$InvokeExpr.eval(Compiler.java:3700)
    ...
  Caused by: java.lang.ClassNotFoundException: javax.xml.bind.DatatypeConverter
          at java.base/java.net.URLClassLoader.findClass(URLClassLoader.java:471)
          at clojure.lang.DynamicClassLoader.findClass(DynamicClassLoader.java:69)
    ...

According to https://stackoverflow.com/a/43574427/80851, a specific set
of modules is missing from recent versions of Java.

This commit pins clojure image to an older tag, e.g. clojure:lein-2.8.1 to fix
building the image.